backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pydantic>=2.10
aiosqlite>=0.20
httpx>=0.28
openai>=1.60
tavily-python>=0.5
numpy>=2.0
python-multipart>=0.0.20
python-dotenv>=1.0
pytest>=8.3
pytest-asyncio>=0.25
